+++ database/migrations/2021_06_16_171442_create_contact_table.php
@@ -14,8 +14,8 @@ class CreateContactTable extends Migration
     public function up()
     {
         Schema::create('contact', function (Blueprint $table) {
-            $table->integer('id');
-            $table->integer('parent_id');
+            $table->bigInteger('id');
+            $table->bigInteger('parent_id');
             $table->string('name');
             $table->timestamps();
         });
